pyproject.toml
[project]
name = "autoline-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "fastapi>=0.116.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
